Separate code into individual files
